Actor's pass along sorted targets.
Added functionality to pass along the corresponding target to an actors
input.

Some of the functions now return an array of size three that hold an
actor's target as it get sorted so it can be passed up to another actor.

Wrote a function to return the target from the size three array since
tokens can't be indexed as part of an expression.